Refactor file service traits and implement stub for testing
- Renamed `FileRemoveService` to `FsRemoveService` for consistency.
- Updated trait definitions to include `Send + Sync` bounds where necessary.
- Introduced a `Stub` implementation for various services to facilitate testing.
- Replaced instances of `MockInfrastructure` with `Stub` in tests across multiple files.
- Added a snapshot test for `usage_prompt` in `tool_service.rs`.
- Created a new snapshot file for tool service usage prompts.
- Ensured all file system related tests utilize the new stub implementation for better isolation and control.
