+++ crates/forge_services/src/infra.rs
@@ -24,7 +24,7 @@ pub trait FsWriteService: Send + Sync {
 }
 
 #[async_trait::async_trait]
-pub trait FileRemoveService: Send + Sync {
+pub trait FsRemoveService: Send + Sync {
     /// Removes a file at the specified path.
     async fn remove(&self, path: &Path) -> anyhow::Result<()>;
 }
@@ -36,7 +36,7 @@ pub trait FsMetaService: Send + Sync {
 }
 
 #[async_trait::async_trait]
-pub trait FsCreateDirsService {
+pub trait FsCreateDirsService: Send + Sync {
     async fn create_dirs(&self, path: &Path) -> anyhow::Result<()>;
 }
 
@@ -76,7 +76,7 @@ pub trait Infrastructure: Send + Sync + Clone + 'static {
     type EnvironmentService: EnvironmentService;
     type FsMetaService: FsMetaService;
     type FsReadService: FsReadService;
-    type FsRemoveService: FileRemoveService;
+    type FsRemoveService: FsRemoveService;
     type FsSnapshotService: FsSnapshotService;
     type FsWriteService: FsWriteService;
     type FsCreateDirsService: FsCreateDirsService;
@@ -92,4 +92,163 @@ pub trait Infrastructure: Send + Sync + Clone + 'static {
     fn create_dirs_service(&self) -> &Self::FsCreateDirsService;
     fn command_executor_service(&self) -> &Self::CommandExecutorService;
     fn inquire_service(&self) -> &Self::InquireService;
+}
+
+#[cfg(test)]
+pub mod stub {
+    use std::sync::Arc;
+
+    use forge_domain::{Environment, EnvironmentService, Provider};
+
+    use super::*;
+
+    impl Default for Stub {
+        fn default() -> Self {
+            Stub {
+                env: Environment {
+                    os: std::env::consts::OS.to_string(),
+                    cwd: std::env::current_dir().unwrap_or_default(),
+                    home: Some("/".into()),
+                    shell: if cfg!(windows) {
+                        "cmd.exe".to_string()
+                    } else {
+                        "/bin/sh".to_string()
+                    },
+                    base_path: PathBuf::new(),
+                    pid: std::process::id(),
+                    provider: Provider::anthropic("test-key"),
+                    retry_config: Default::default(),
+                },
+            }
+        }
+    }
+
+    #[derive(Clone)]
+    pub struct Stub {
+        env: Environment,
+    }
+
+    #[async_trait::async_trait]
+    impl EnvironmentService for Stub {
+        fn get_environment(&self) -> Environment {
+            self.env.clone()
+        }
+    }
+
+    #[async_trait::async_trait]
+    impl FsReadService for Stub {
+        async fn read(&self, _path: &Path) -> anyhow::Result<Bytes> {
+            unimplemented!()
+        }
+    }
+
+    #[async_trait::async_trait]
+    impl FsWriteService for Stub {
+        async fn write(&self, _: &Path, _: Bytes) -> anyhow::Result<()> {
+            unimplemented!()
+        }
+    }
+
+    #[async_trait::async_trait]
+    impl FsSnapshotService for Stub {
+        async fn create_snapshot(&self, _: &Path) -> anyhow::Result<Snapshot> {
+            unimplemented!()
+        }
+
+        async fn undo_snapshot(&self, _: &Path) -> anyhow::Result<()> {
+            Ok(())
+        }
+    }
+
+    #[async_trait::async_trait]
+    impl FsMetaService for Stub {
+        async fn is_file(&self, _: &Path) -> anyhow::Result<bool> {
+            unimplemented!()
+        }
+
+        async fn exists(&self, _: &Path) -> anyhow::Result<bool> {
+            unimplemented!()
+        }
+    }
+
+    #[async_trait::async_trait]
+    impl FsRemoveService for Stub {
+        async fn remove(&self, _: &Path) -> anyhow::Result<()> {
+            unimplemented!()
+        }
+    }
+
+    #[async_trait::async_trait]
+    impl FsCreateDirsService for Stub {
+        async fn create_dirs(&self, _: &Path) -> anyhow::Result<()> {
+            unimplemented!()
+        }
+    }
+
+    #[async_trait::async_trait]
+    impl CommandExecutorService for Stub {
+        async fn execute_command(&self, _: String, _: PathBuf) -> anyhow::Result<CommandOutput> {
+            unimplemented!()
+        }
+    }
+
+    #[async_trait::async_trait]
+    impl InquireService for Stub {
+        async fn select_one(&self, _: &str, _: Vec<String>) -> anyhow::Result<String> {
+            unimplemented!()
+        }
+
+        async fn select_many(&self, _: &str, _: Vec<String>) -> anyhow::Result<Vec<String>> {
+            unimplemented!()
+        }
+    }
+
+    #[async_trait::async_trait]
+    impl Infrastructure for Stub {
+        type EnvironmentService = Stub;
+        type FsReadService = Stub;
+        type FsWriteService = Stub;
+        type FsRemoveService = Stub;
+        type FsMetaService = Stub;
+        type FsSnapshotService = Stub;
+        type FsCreateDirsService = Stub;
+        type CommandExecutorService = Stub;
+        type InquireService = Stub;
+
+        fn environment_service(&self) -> &Self::EnvironmentService {
+            self
+        }
+
+        fn file_read_service(&self) -> &Self::FsReadService {
+            self
+        }
+
+        fn file_write_service(&self) -> &Self::FsWriteService {
+            self
+        }
+
+        fn file_meta_service(&self) -> &Self::FsMetaService {
+            self
+        }
+
+        fn file_snapshot_service(&self) -> &Self::FsSnapshotService {
+            self
+        }
+
+        fn file_remove_service(&self) -> &Self::FsRemoveService {
+            self
+        }
+
+        fn create_dirs_service(&self) -> &Self::FsCreateDirsService {
+            self
+        }
+
+        fn command_executor_service(&self) -> &Self::CommandExecutorService {
+            self
+        }
+
+        fn inquire_service(&self) -> &Self::InquireService {
+            self
+        }
+    }
 }
