+++ crates/forge_services/src/tools/registry.rs
@@ -40,173 +40,11 @@ impl<F: Infrastructure> ToolRegistry<F> {
 
 #[cfg(test)]
 pub mod tests {
-    use std::path::{Path, PathBuf};
 
-    use bytes::Bytes;
-    use forge_domain::{CommandOutput, Environment, EnvironmentService, Provider, ToolUsagePrompt};
-    use forge_snaps::Snapshot;
+    use forge_domain::ToolUsagePrompt;
     use insta::assert_snapshot;
 
     use super::*;
-    use crate::{
-        CommandExecutorService, FileRemoveService, FsCreateDirsService, FsMetaService,
-        FsReadService, FsSnapshotService, FsWriteService, InquireService,
-    };
-
-    /// Create a default test environment
-    fn stub() -> Stub {
-        Stub {
-            env: Environment {
-                os: std::env::consts::OS.to_string(),
-                cwd: std::env::current_dir().unwrap_or_default(),
-                home: Some("/".into()),
-                shell: if cfg!(windows) {
-                    "cmd.exe".to_string()
-                } else {
-                    "/bin/sh".to_string()
-                },
-                base_path: PathBuf::new(),
-                pid: std::process::id(),
-                provider: Provider::anthropic("test-key"),
-                retry_config: Default::default(),
-            },
-        }
-    }
-
-    impl Default for Stub {
-        fn default() -> Self {
-            stub()
-        }
-    }
-
-    #[derive(Clone)]
-    pub struct Stub {
-        env: Environment,
-    }
-
-    #[async_trait::async_trait]
-    impl EnvironmentService for Stub {
-        fn get_environment(&self) -> Environment {
-            self.env.clone()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl FsReadService for Stub {
-        async fn read(&self, _path: &Path) -> anyhow::Result<Bytes> {
-            unimplemented!()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl FsWriteService for Stub {
-        async fn write(&self, _: &Path, _: Bytes) -> anyhow::Result<()> {
-            unimplemented!()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl FsSnapshotService for Stub {
-        async fn create_snapshot(&self, _: &Path) -> anyhow::Result<Snapshot> {
-            unimplemented!()
-        }
-
-        async fn undo_snapshot(&self, _: &Path) -> anyhow::Result<()> {
-            Ok(())
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl FsMetaService for Stub {
-        async fn is_file(&self, _: &Path) -> anyhow::Result<bool> {
-            unimplemented!()
-        }
-
-        async fn exists(&self, _: &Path) -> anyhow::Result<bool> {
-            unimplemented!()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl FileRemoveService for Stub {
-        async fn remove(&self, _: &Path) -> anyhow::Result<()> {
-            unimplemented!()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl FsCreateDirsService for Stub {
-        async fn create_dirs(&self, _: &Path) -> anyhow::Result<()> {
-            unimplemented!()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl CommandExecutorService for Stub {
-        async fn execute_command(&self, _: String, _: PathBuf) -> anyhow::Result<CommandOutput> {
-            unimplemented!()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl InquireService for Stub {
-        async fn select_one(&self, _: &str, _: Vec<String>) -> anyhow::Result<String> {
-            unimplemented!()
-        }
-
-        async fn select_many(&self, _: &str, _: Vec<String>) -> anyhow::Result<Vec<String>> {
-            unimplemented!()
-        }
-    }
-
-    #[async_trait::async_trait]
-    impl Infrastructure for Stub {
-        type EnvironmentService = Stub;
-        type FsReadService = Stub;
-        type FsWriteService = Stub;
-        type FsRemoveService = Stub;
-        type FsMetaService = Stub;
-        type FsSnapshotService = Stub;
-        type FsCreateDirsService = Stub;
-        type CommandExecutorService = Stub;
-        type InquireService = Stub;
-
-        fn environment_service(&self) -> &Self::EnvironmentService {
-            self
-        }
-
-        fn file_read_service(&self) -> &Self::FsReadService {
-            self
-        }
-
-        fn file_write_service(&self) -> &Self::FsWriteService {
-            self
-        }
-
-        fn file_meta_service(&self) -> &Self::FsMetaService {
-            self
-        }
-
-        fn file_snapshot_service(&self) -> &Self::FsSnapshotService {
-            self
-        }
-
-        fn file_remove_service(&self) -> &Self::FsRemoveService {
-            self
-        }
-
-        fn create_dirs_service(&self) -> &Self::FsCreateDirsService {
-            self
-        }
-
-        fn command_executor_service(&self) -> &Self::CommandExecutorService {
-            self
-        }
-
-        fn inquire_service(&self) -> &Self::InquireService {
-            self
-        }
-    }
 
     #[test]
     fn test_tool_description_length() {
@@ -215,7 +53,7 @@ pub mod tests {
         println!("\nTool description lengths:");
 
         let mut any_exceeded = false;
-        let stub = Arc::new(stub());
+        let stub = Arc::new(crate::infra::stub::Stub::default());
         let registry = ToolRegistry::new(stub.clone());
         for tool in registry.tools() {
             let desc_len = tool.definition.description.len();
@@ -244,7 +82,7 @@ pub mod tests {
 
     #[test]
     fn test_tool_usage() {
-        let stub = Arc::new(stub());
+        let stub = Arc::new(crate::infra::stub::Stub::default());
         let registry = ToolRegistry::new(stub.clone());
         let tools = registry
             .tools()
